Add files via upload
The algorithm (for a postgraduate test)  simulates the course of an epidemic in a population according to the SEIR model, solving a system of ordinary differential equations and visualising the evolution over time of the different categories of individuals (Susceptible, Exposed, Infected, Healed). It is also shown how different initial configurations influence the behaviour of the model.